[Impeller] dont use blend shader for simple drawAtlas calls. (#55420)
When making a drawAtlas call with no colors, use the faster shader that does not apply blending.

Now that we're switched to experimental canvas, we can remove the copying of the display list data. The draw can directly refer to pointers from the display list structure, as the draw is processed during DL dispatching.